requirements.txt
django
djangorestframework
django-filter
django-cors-headers
Pillow
drf_yasg
